+++ src/index.ts
@@ -1,27 +1,30 @@
-import { SentenceFactory } from './factory';
-import { Sentence } from './sentence';
+// eslint-disable-next-line no-unused-vars
+import { Template, Vocabulary, Options, Configuration, StringResolvable, WeightedEntry } from '../types';
+import { Sentence, SentenceFactory } from './sentence';
+import defaults from './defaults';
 
-const Facade = new SentenceFactory();
+const Facade = new SentenceFactory(defaults.templates, defaults.vocabulary, {});
 
 export const createSentence = (
   templates?: Template[] | Template,
   vocabulary?: Vocabulary,
-  options?: MaybeOptions,
+  options?: Options,
 ) => Facade.createSentence(templates, vocabulary, options);
+
 export const configure = (config: Configuration) => {
   Facade.configure(config);
 };
-export const addDefaultOptions = (options: MaybeOptions) => {
+
+export const addDefaultOptions = (options: Options) => {
   Facade.addDefaultOptions(options);
 };
+
 export const addDefaultTemplates = (...templates: Template[]) => {
   Facade.addDefaultTemplates(...templates);
 };
+
 export const addDefaultVocabulary = (vocab: Vocabulary) => {
   Facade.addDefaultVocabulary(vocab);
 };
-export const restoreDefaults = () => {
-  Facade.restoreDefaults();
-};
 
-export { Sentence, SentenceFactory };
+export { Sentence, SentenceFactory, StringResolvable, WeightedEntry, Template, Vocabulary, Options, Configuration };