+++ src/sentence/Sentence.ts
@@ -1,24 +1,31 @@
-const defaultOptions: Options = {
-  capitalize: true,
-  forceNewSentence: false,
-  placeholderNotation: {
-    start: '{',
-    end: '}'
-  },
-  preservePlaceholderNotation: false,
-};
+/* eslint-disable no-unused-vars */
+import {
+  DefinitelyOptions,
+  WeightedTemplate,
+  WeightedVocabulary,
+  Template,
+  Vocabulary,
+  Options,
+  Configuration,
+  WeightedEntry,
+  StringResolvable,
+  PlaceholderNotation
+} from '../../types';
+/* eslint-enable no-unused-vars */
+
+import defaults from '../defaults';
 
 export class Sentence {
   #templates: WeightedTemplate[] = [];
   #vocabulary: WeightedVocabulary = {};
-  #options: Options = defaultOptions;
+  #options: DefinitelyOptions = defaults.options;
 
   public value: string = '';
 
   constructor(
     templates: Template[] | Template,
     vocabulary: Vocabulary,
-    options?: MaybeOptions,
+    options?: Options,
   ) {
     this.configure({
       options: options,
@@ -47,14 +54,14 @@ export class Sentence {
   }
 
   public restoreDefaultOptions(): void {
-    this.options = defaultOptions;
+    this.options = defaults.options;
   }
 
-  public get options(): MaybeOptions {
+  public get options(): Options {
     return this.#options;
   }
 
-  public set options(options: MaybeOptions) {
+  public set options(options: Options) {
     const { placeholderNotation } = options;
     if (placeholderNotation && typeof placeholderNotation == 'string') {
       options.placeholderNotation = this.parsePlaceholderNotation(placeholderNotation);
@@ -62,7 +69,7 @@ export class Sentence {
     this.#options = {
       ...this.options,
       ...options,
-    } as Options;
+    } as DefinitelyOptions;
   }
 
   public get templates(): Template[] {
@@ -145,7 +152,7 @@ export class Sentence {
    * 'This is {a-adjective} example.' => ['{a-adjective}']
    */
   private findPlaceholders(template: StringResolvable): RegExpMatchArray | null {
-    const { placeholderNotation } = this.options as Options;
+    const { placeholderNotation } = this.options as DefinitelyOptions;
     const { start, end } = placeholderNotation;
     const regex = new RegExp(`([${start}]+(\\s*([a-z-0-9])*,?\\s*)*[${end}]+)`, 'gi');
     return (typeof template === 'string') ? template.match(regex) : template().match(regex);
@@ -159,7 +166,7 @@ export class Sentence {
     const alternatives = this.resolveAlternatives(placeholder);
     const chosenWord = shouldCapitalize ? capitalize(this.pickRandomEntryByWeight(alternatives)) : this.pickRandomEntryByWeight(alternatives);
 
-    const { placeholderNotation, preservePlaceholderNotation } = this.options as Options;
+    const { placeholderNotation, preservePlaceholderNotation } = this.options as DefinitelyOptions;
     if (preservePlaceholderNotation) {
       const { start, end } = placeholderNotation;
       return `${start}${chosenWord}${end}`;
@@ -224,7 +231,7 @@ export class Sentence {
    * '{a-adjective, a-curse, verb}' => ['a-adjective', 'a-curse', 'verb']
    */
   private findKeys(placeholder: string): string[] {
-    const { placeholderNotation } = this.options as Options;
+    const { placeholderNotation } = this.options as DefinitelyOptions;
     const { start, end } = placeholderNotation;
     return placeholder.replace(new RegExp(`${start}|${end}|\\s`, 'g'), '').split(',');
   }