+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [Unreleased] 1.0.0
+> TBD
+### Breaking :boom:
+* **Changed** various types and interfaces
+* **Removed** method `restoreDefaults` from `SentenceFactory`
+### Features :tada:
+* **Added** needlessly complex esm build process
+* **Added** automatic type declarations to compile
+* **Added** following types to exports:
+  * `Template` `Vocabulary` `Options` `Configuration` `StringResolvable` `WeightedEntry`
+
 ## [0.6.1](../../tree/v0.6.1)
 ### Bugfix :bug:
 * **Fix** error where adding `a-` or `-s` to a placeholder would break during runtime